Modified Part 1 code

+++ Part1_encryption.py
@@ -1,11 +1,13 @@
 from imports_and_libraries import *
 from file_encryption import *
+f = open('user/word_indices.json')
+word_indices = json.load(f)
 
-def generate_index_string(number_of_words,file_num,word_indices):
+def generate_index_string(file_name,number_of_words,file_num,word_indices):
     I = [0 for i in range(number_of_words)]
 
     
-    file = open("user/{}.txt".format(file_num),'r')
+    file = open("user/{}.txt".format(file_name[:-4]),'r')
     text = file.read().lower()
     text_list = word_tokenize(text)
     for i in range(number_of_words):
@@ -27,26 +29,55 @@ def generate_index_string(number_of_words,file_num,word_indices):
     
     return final_string
 
-def pass_to_server(number_of_files,word_indices):
-    for i in range(number_of_files):
-        file_name = '{}.txt'.format(i + 1)
-        data = generate_index_string(NUMBER_OF_WORDS,i + 1,word_indices)
-        encrypt_file(file_name)
-        # print(data)
-        file_write = open("server/"+file_name[:-4]+".enc","a")
-        file_write.write(data)
-        file_write.close()
+def pass_to_server(file_name,word_indices):
+    
+    fname = "user/file_indices.json"
+    with open(fname) as file_indices:
+        file_indices_json = json.load(file_indices)
+
+    file_indices_json["file_index"][file_name] = file_indices_json["total"]
+    file_indices_json["total"] += 1
+    index = file_indices_json["total"]
+
+    with open(fname, mode='w') as f:
+        f.write(json.dumps(file_indices_json, indent=2))
+
+    with open("server/file_indices.json", mode='w') as f:
+        f.write(json.dumps(file_indices_json, indent=2))
+    
+    data = generate_index_string(file_name,NUMBER_OF_WORDS,index,word_indices)
+    encrypt_file(file_name)
+    # print(data)
+    file_write = open("server/"+file_name[:-4]+".enc","a")
+    file_write.write(data)
+    file_write.close()
 
 if __name__ == "__main__":
-    s = generate_pad(t)
-    file_write = open("user/s.txt","w")
-    file_write.write(s)
-    file_write.close()
-    r = generate_pad(t)
-    file_write = open("user/r.txt","w")
-    file_write.write(r)
-    file_write.close()
+    if not os.path.isfile("user/s.txt"):
+        s = generate_pad_random(t)
+        file_write = open("user/s.txt","w")
+        file_write.write(s)
+        file_write.close()
+    else:
+        file_read = open("user/s.txt","r")
+        s = file_read.read()
+        file_read.close()
+    
+    if not os.path.isfile("user/r.txt"):
+        r = generate_pad_random(t)
+        file_write = open("user/r.txt","w")
+        file_write.write(r)
+        file_write.close()
+    else:
+        file_read = open("user/r.txt","r")
+        r = file_read.read()
+        file_read.close()
 
-    pass_to_server(NUMBER_OF_FILES ,word_indices)
+    print("Enter File Name for Encryption: ",end="")
+    file_name = input()
+    if os.path.isfile("user/{}".format(file_name)):
+        pass_to_server(file_name ,word_indices)
+    else:
+        print("File Not Found")
